Move biconsumer to common4j
