+++ common/src/main/java/com/microsoft/identity/common/internal/broker/BoundServiceClient.java
@@ -36,7 +36,7 @@ import androidx.annotation.Nullable;
 
 import com.microsoft.identity.common.exception.BrokerCommunicationException;
 import com.microsoft.identity.common.internal.broker.ipc.BrokerOperationBundle;
-import com.microsoft.identity.common.internal.result.ResultFuture;
+import com.microsoft.identity.common.java.util.ResultFuture;
 import com.microsoft.identity.common.logging.Logger;
 
 import java.util.List;
