+++ common4j/src/main/com/microsoft/identity/common/java/util/ResultFuture.java
@@ -1,3 +1,4 @@
+
 //  Copyright (c) Microsoft Corporation.
 //  All rights reserved.
 //
@@ -20,12 +21,11 @@
 //  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 //  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 //  THE SOFTWARE.
-package com.microsoft.identity.common.internal.result;
 
-import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
+package com.microsoft.identity.common.java.util;
 
-import com.microsoft.identity.common.internal.util.BiConsumer;
+import edu.umd.cs.findbugs.annotations.Nullable;
+import lombok.NonNull;
 
 import java.util.ArrayList;
 import java.util.List;
